Add config.py. Refact: modul stellarisburger -> class UserAccount
