+++ config.py
@@ -1,0 +1,1 @@
+URL = 'https://stellarburgers.nomoreparties.site'
